mcp-servers/requirements.txt
# MCP Server Requirements
# Install with: pip install -r requirements.txt

# FastMCP - Simplified MCP server framework
fastmcp>=0.3.0

# Note: The MCP servers use the same dependencies as the main project
# (selenium, pandas, openpyxl, etc.) but those are already installed
# in the conda environment. This file only lists MCP-specific dependencies.
